move connection to gateways where it belongs
